refactor(negative): reorganize keywords
