+++ e2e/keywords/common.resource
@@ -4,7 +4,9 @@ Documentation       Common keywords
 Library             ../libs/keywords/common_keywords.py
 Library             ../libs/keywords/stress_keywords.py
 Library             ../libs/keywords/volume_keywords.py
-Library             ../libs/keywords/recurring_job_keywords.py
+Library             ../libs/keywords/recurringjob_keywords.py
+Library             ../libs/keywords/deployment_keywords.py
+Library             ../libs/keywords/statefulset_keywords.py
 Library             ../libs/keywords/workload_keywords.py
 Library             ../libs/keywords/network_keywords.py
 
@@ -31,7 +33,7 @@ Cleanup test resources
     cleanup_control_plane_network_latency
     cleanup_node_exec
     cleanup_stress_helper
-    cleanup_recurring_jobs    ${volume_list}
+    cleanup_recurringjobs    ${volume_list}
     cleanup_volumes    ${volume_list}
     cleanup_deployments    ${deployment_list}
     cleanup_statefulsets    ${statefulset_list}
