+++ e2e/keywords/kubelet.resource
@@ -11,5 +11,5 @@ Library             ../libs/keywords/volume_keywords.py
 *** Keywords ***
 Stop volume node kubelet of statefulset ${idx} for ${stop_time_in_sec} seconds
     ${volume_name} =    get_workload_volume_name    ${statefulset_list}[${idx}]
-    ${node_name} =    get_volume_node    ${volume_name}
+    ${node_name} =    get_replica_node_attached_to_volume    ${volume_name}
     restart_kubelet    ${node_name}    ${stop_time_in_sec}
