--- e2e/libs/persistent_volume_claim/__init__.py
@@ -1,1 +1,0 @@
-from persistent_volume_claim.persistent_volume_claim import PersistentVolumeClaim
