+++ e2e/libs/persistentvolumeclaim/__init__.py
@@ -1,0 +1,1 @@
+from persistentvolumeclaim.persistentvolumeclaim import PersistentVolumeClaim
