--- e2e/libs/recurring_job/__init__.py
@@ -1,1 +1,0 @@
-from recurring_job.recurring_job import RecurringJob
