+++ e2e/libs/recurringjob/__init__.py
@@ -1,0 +1,1 @@
+from recurringjob.recurringjob import RecurringJob
