+++ e2e/tests/cluster_restart.robot
@@ -1,8 +1,10 @@
 *** Settings ***
 Documentation    Negative Test Cases
-Resource    ../keywords/workload.resource
-Resource    ../keywords/node.resource
 Resource    ../keywords/common.resource
+Resource    ../keywords/deployment.resource
+Resource    ../keywords/node.resource
+Resource    ../keywords/statefulset.resource
+Resource    ../keywords/workload.resource
 
 Test Setup    Set test environment
 Test Teardown    Cleanup test resources
