+++ e2e/tests/node_reboot.robot
@@ -1,9 +1,12 @@
 *** Settings ***
 Documentation    Negative Test Cases
-Resource    ../keywords/workload.resource
-Resource    ../keywords/volume.resource
-Resource    ../keywords/node.resource
 Resource    ../keywords/common.resource
+Resource    ../keywords/deployment.resource
+Resource    ../keywords/node.resource
+Resource    ../keywords/recurringjob.resource
+Resource    ../keywords/statefulset.resource
+Resource    ../keywords/volume.resource
+Resource    ../keywords/workload.resource
 
 Test Setup    Set test environment
 Test Teardown    Cleanup test resources
@@ -151,4 +154,38 @@ Power Off Volume Node For More Than Pod Eviction Timeout While Workload Heavy Wr
         And Wait for statefulset 0 stable
 
         Then Check statefulset 0 works
+    END
+
+Reboot Volume Node While Heavy Writing And Recurring Jobs Exist
+    Given Create volume 0 with 2 GB and 1 replicas
+    And Create volume 1 with 2 GB and 3 replicas
+    And Keep writing data to volume 0
+    And Keep Writing data to volume 1
+    And Create snapshot and backup recurringjob for volume 0
+    And Create snapshot and backup recurringjob for volume 1
+
+    FOR    ${i}    IN RANGE    ${LOOP_COUNT}
+        When Reboot volume 0 volume node
+
+        Then Check recurringjobs for volume 0 work
+        And Check recurringjobs for volume 1 work
+        And Check volume 0 works
+        And Check volume 1 works
+    END
+
+Reboot Replica Node While Heavy Writing And Recurring Jobs Exist
+    Given Create volume 0 with 2 GB and 1 replicas
+    And Create volume 1 with 2 GB and 3 replicas
+    And Keep Writing data to volume 0
+    And Keep Writing data to volume 1
+    And Create snapshot and backup recurringjob for volume 0
+    And Create snapshot and backup recurringjob for volume 1
+
+    FOR    ${i}    IN RANGE    ${LOOP_COUNT}
+        When Reboot volume 1 replica node
+
+        Then Check recurringjobs for volume 0 work
+        And Check recurringjobs for volume 1 work
+        And Check volume 0 works
+        And Check volume 1 works
     END
