+++ e2e/tests/replica_rebuilding.robot
@@ -1,8 +1,8 @@
 *** Settings ***
 Documentation    Negative Test Cases
-Resource    ../keywords/volume.resource
-Resource    ../keywords/node.resource
 Resource    ../keywords/common.resource
+Resource    ../keywords/node.resource
+Resource    ../keywords/volume.resource
 
 Test Setup    Set test environment
 Test Teardown    Cleanup test resources
@@ -13,6 +13,19 @@ ${RETRY_COUNT}    300
 ${RETRY_INTERVAL}    1
 
 *** Test Cases ***
+Delete Replica While Replica Rebuilding
+    Given Create a volume with 2 GB and 3 replicas
+    And Write data to the volume
+
+    FOR    ${i}    IN RANGE    ${LOOP_COUNT}
+        When Delete replica 0 to trigger replica 0 rebuilding
+        And During replica 0 rebuilding, delete replica 1
+        And Wait until replica 0 rebuilt, delete replica 2
+
+        Then Check data is intact
+        And Wait until all replicas rebuilt
+    END
+
 Reboot Volume Node While Replica Rebuilding
     Given Create a volume with 5 GB and 3 replicas
     And Write data to the volume
